Add Linux generic support
The new Linux generic support is actually the original code from
lstools_ubuntu.

Now MbedLsToolsUbuntu inherits from MbedLsToolsLinuxGeneric (to allow
Ubuntu-specific changes in the future).
